// mobile/src/theme.d.ts
export declare const colors: {
    navy: string;
    navyCard: string;
    navyCardAlt: string;
    navyBorder: string;
    lightBg: string;
    white: string;
    yellow: string;
    textLight: string;
    textMutedLight: string;
    textDark: string;
    textMutedDark: string;
    green: string;
    greenSoft: string;
    red: string;
    redSoft: string;
    orange: string;
    orangeSoft: string;
    purple: string;
    blue: string;
    border: string;
};
export declare const radius: {
    sm: number;
    md: number;
    lg: number;
    xl: number;
};
export declare const spacing: {
    xs: number;
    sm: number;
    md: number;
    lg: number;
    xl: number;
};